+++ src/header.hpp
@@ -82,4 +82,4 @@ bool is_valid_key(char parameter_key);
 #include "model/Network.hpp"
 #include "model/Model.hpp"
 #include "model/Metrics.hpp"
-#include "Worker.hpp"
+#include "ModelRunner.hpp"
